requirements.txt
# 大模型接口调用
requests>=2.28.0
openai>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.8

# 数据分析相关
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.10.0

# 数据可视化
plotly>=5.14.0
dash>=2.9.0

# 数据处理和工作流
dask>=2023.5.0
prefect>=2.10.0

# 数据库连接
sqlalchemy>=2.0.0
pymysql>=1.0.0
psycopg2-binary>=2.9.5

# API服务
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 开发工具
python-dotenv>=1.0.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# 测试框架
pytest>=7.3.1
pytest-cov>=4.0.0
httpx>=0.24.0

# 工具类
tqdm>=4.65.0